+++ PsycheAssistant/composeApp/src/commonMain/kotlin/Composable/Layout/PsycheAssistantApp.kt
@@ -1,15 +1,21 @@
 package org.psyche.assistant.Composable.Layout
 
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.List
-import androidx.compose.material.icons.filled.Done
-import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material.icons.twotone.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import org.jetbrains.compose.resources.painterResource
+
 import org.jetbrains.compose.resources.stringResource
 import org.psyche.assistant.Composable.Common.ErrorScreen
 import org.psyche.assistant.Composable.Common.LoadingScreen
@@ -17,7 +23,7 @@ import org.psyche.assistant.Composable.LocalAuthToken
 import org.psyche.assistant.Composable.LocalGroup
 import org.psyche.assistant.Composable.LocalUser
 import org.psyche.assistant.Composable.Main.MainPage
-import org.psyche.assistant.Composable.Settings.SettingsPage
+import org.psyche.assistant.Composable.Settings.AccountManagementPage
 import org.psyche.assistant.Composable.Survey.SurveyPage
 import org.psyche.assistant.Controller.GroupController
 import org.psyche.assistant.Controller.UserController
@@ -28,21 +34,17 @@ import org.psyche.assistant.Storage.AuthStorage
 import psycheassistant.composeapp.generated.resources.Res
 import psycheassistant.composeapp.generated.resources.*
 
-
-/**
- * Composable that controls the screen, adding a nav-bar at the top and a switch mechanism for navigating between them.
- * I.e. when one item is clicked, the "currentScreen" variable changes, and the appropriate Composable is called.
- */
 @Composable
 fun PsycheAssistantApp() {
     var currentScreen by remember { mutableStateOf("main") }
     var authToken = remember { mutableStateOf(AuthStorage.getAuthToken()) }
     val userState = remember { mutableStateOf<User?>(null) }
-    val groupState = remember { mutableStateOf<Group?>(null)}
+    val groupState = remember { mutableStateOf<Group?>(null) }
 
-    var isLoading by remember { mutableStateOf(true)}
-    var isError by remember { mutableStateOf(false)}
-    var errorMessage by remember { mutableStateOf("")}
+    var isLoading by remember { mutableStateOf(true) }
+    var isError by remember { mutableStateOf(false) }
+    var displayError by remember { mutableStateOf(false) }
+    var errorMessage by remember { mutableStateOf("") }
 
     LaunchedEffect(authToken.value) {
         if (authToken.value != null) {
@@ -56,7 +58,8 @@ fun PsycheAssistantApp() {
             } catch (e: Exception) {
                 isLoading = false
                 isError = true
-                errorMessage = e.message.toString();
+                displayError = true
+                errorMessage = e.message.toString()
             }
         } else {
             isLoading = false
@@ -64,71 +67,111 @@ fun PsycheAssistantApp() {
     }
 
     MaterialTheme(
-        /**
-         * Wrap the entire contents in the custom color theme.
-         */
         colors = CustomTheme.psycheColors()
     ) {
-        // Get the GlobalState value and pass it along to the compositions, allowing for updating the value across screens.
         CompositionLocalProvider(
             LocalAuthToken provides authToken,
             LocalUser provides userState,
             LocalGroup provides groupState
         ) {
             Scaffold(
-                /**
-                 * Add a top bar for navigation, passing new values to "currentScreen" when selected, controlling the inner contents.
-                 */
                 topBar = {
-                    TopAppBar {
-                        BottomNavigationItem(
-                            icon = { Icon(Icons.Default.Home, contentDescription = null) },
-                            label = { Text(stringResource(Res.string.home)) },
-                            selected = currentScreen == "main",
-                            onClick = { currentScreen = "main" }
-                        )
-                        BottomNavigationItem(
-                            icon = { Icon(Icons.AutoMirrored.Filled.List, contentDescription = null) },
-                            label = { Text(stringResource(Res.string.survey)) },
-                            selected = currentScreen == "survey",
-                            onClick = { currentScreen = "survey" }
-                        )
-                        BottomNavigationItem(
-                            icon = { Icon(Icons.Default.Done, contentDescription = null) },
-                            label = { Text(stringResource(Res.string.settings)) },
-                            selected = currentScreen == "settings",
-                            onClick = { currentScreen = "settings" }
+                    Box(
+                        modifier = Modifier
+                            .clickable { currentScreen = "main" } // Make the entire TopAppBar clickable
+                            .fillMaxWidth()
+                    ) {
+                        TopAppBar(
+                            title = {
+                                Text(
+                                    text = stringResource(Res.string.home),
+                                    maxLines = 1,
+                                    overflow = TextOverflow.Ellipsis,
+                                )
+                            },
+                            navigationIcon = {
+                                val icon: Painter = painterResource(Res.drawable.psyche_assistant_icon)
+                                Icon(
+                                    painter = icon,
+                                    contentDescription = stringResource(Res.string.home),
+                                    tint = Color.Unspecified
+                                )
+                            },
                         )
                     }
                 },
-
-                /**
-                 * Add content, changing depending on the currentScreen value.
-                 */
+                bottomBar = {
+                    BottomNavigation {
+                        BottomNavigationItem(
+                            icon = { Icon(Icons.Filled.List, contentDescription = null) },
+                            label = { Text(stringResource(Res.string.survey), style = TextStyle(fontSize = 11.sp), maxLines = 1, overflow = TextOverflow.Ellipsis) },
+                            selected = currentScreen == "survey",
+                            onClick = { currentScreen = "survey" }
+                        )
+                        if (!isError) {
+                            BottomNavigationItem(
+                                icon = { Icon(Icons.TwoTone.CheckCircle, contentDescription = null) },
+                                label = {
+                                    Text(
+                                        stringResource(Res.string.activities),
+                                        style = TextStyle(fontSize = 11.sp),
+                                        maxLines = 1,
+                                        overflow = TextOverflow.Ellipsis
+                                    )
+                                },
+                                selected = currentScreen == "activities",
+                                onClick = { currentScreen = "activities" }
+                            )
+                            BottomNavigationItem(
+                                icon = { Icon(Icons.TwoTone.Settings, contentDescription = null) },
+                                label = {
+                                    Text(
+                                        stringResource(Res.string.settings),
+                                        style = TextStyle(fontSize = 11.sp),
+                                        maxLines = 1,
+                                        overflow = TextOverflow.Ellipsis
+                                    )
+                                },
+                                selected = currentScreen == "settings",
+                                onClick = { currentScreen = "settings" }
+                            )
+                        }
+                    }
+                },
                 content = { innerPadding ->
                     if (isLoading) {
                         LoadingScreen()
-                    } else if (isError) {
-                        ErrorScreen(message = errorMessage)
                     } else {
                         Column(
                             modifier = Modifier
                                 .padding(innerPadding)
                                 .fillMaxSize()
-                                .verticalScroll(rememberScrollState())
                         ) {
                             when (currentScreen) {
                                 "main" -> MainPage()
                                 "survey" -> SurveyPage(onBack = { currentScreen = "main" })
-                                "settings" -> SettingsPage()
+                                "settings" -> AccountManagementPage()
                             }
                         }
                     }
                 }
             )
+
+            // Error dialog
+            if (displayError) {
+                AlertDialog(
+                    onDismissRequest = { displayError = false }, // Hide the dialog when the user clicks outside
+                    title = { Text(stringResource(Res.string.no_connection)) },
+                    text = { Text(errorMessage) },
+                    confirmButton = {
+                        Button(
+                            onClick = { displayError = false } // Hide the dialog when the user clicks "OK"
+                        ) {
+                            Text("OK")
+                        }
+                    }
+                )
+            }
         }
     }
-
 }
-
-