+++ PsycheAssistant/composeApp/src/commonMain/kotlin/Model/Group/Group.kt
@@ -10,5 +10,8 @@ data class Group (
     // Clarify field name since back-end operates with objects, and front-end with IDs
     @SerialName("users")
     val userIds: List<Int>,
+    @SerialName("owner")
+    val ownerId: Int?,
+
     val code: String,
 )