+++ PsycheAssistant/composeApp/src/commonMain/kotlin/Service/GroupService.kt
@@ -4,6 +4,7 @@ import io.ktor.client.*
 import io.ktor.client.request.*
 import io.ktor.client.statement.*
 import io.ktor.http.*
+import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import org.psyche.assistant.Model.Group.Group
 import org.psyche.assistant.Model.Group.GroupRepository
@@ -27,10 +28,21 @@ class GroupService : GroupRepository {
         }
     }
 
-    override suspend fun joinGroup(code: String) {
-        TODO("Not yet implemented")
+    override suspend fun joinGroup(authToken: String, code: String): Group? {
+        val url = ServiceBuilder.url("api/groups/join")
+        val response: HttpResponse = client.post(url) {
+            contentType(ContentType.Application.Json)
+            header(HttpHeaders.Authorization, "Bearer $authToken")
+            setBody(Json.encodeToString(mapOf("code" to code)))
+        }
+        return when (response.status) {
+            HttpStatusCode.OK -> {
+                val responseBody = response.bodyAsText()
+                Json.decodeFromString<Group>(responseBody)
+            }
+            else -> null
+        }
     }
-
     override suspend fun leaveGroup() {
         TODO("Not yet implemented")
     }
@@ -39,6 +51,23 @@ class GroupService : GroupRepository {
         val url = ServiceBuilder.url("api/groups/$id")
         val response: HttpResponse = client.get(url) {
             contentType(ContentType.Application.Json)
+        }
+
+        return when (response.status) {
+            HttpStatusCode.OK -> {
+                val responseBody = response.bodyAsText()
+                Json.decodeFromString<Group>(responseBody)
+            }
+            else -> null
+        }
+    }
+
+    override suspend fun removeUserFromGroup(authToken: String, groupId: Int, userId: Int): Group? {
+        val url = ServiceBuilder.url("api/groups/$groupId/kick")
+        val response: HttpResponse = client.post(url) {
+            contentType(ContentType.Application.Json)
+            header(HttpHeaders.Authorization, "Bearer $authToken")
+            setBody(Json.encodeToString(mapOf("userId" to userId)))
         }
 
         return when (response.status) {