+++ PsycheAssistant/composeApp/src/commonMain/kotlin/Service/UserService.kt
@@ -26,7 +26,13 @@ class UserService : UserRepository {
         }
     }
 
+    /**
+     * Register the user using the supplied credentials. Verify that email and password live up to security standards, before posting to server.
+     */
     override suspend fun registerUser(email: String, password: String): String {
+        validateEmail(email)
+        validatePassword(password)
+
         val url = ServiceBuilder.url("api/users/register")
         val response: HttpResponse = client.post(url) {
             contentType(ContentType.Application.FormUrlEncoded)
@@ -34,7 +40,7 @@ class UserService : UserRepository {
         }
         return when (response.status) {
             HttpStatusCode.Created -> response.bodyAsText()
-            else -> throw Exception("Failed to register user: ${response.status}")
+            else -> throw Exception(response.bodyAsText())
         }
     }
 
@@ -49,11 +55,43 @@ class UserService : UserRepository {
             AuthStorage.saveAuthToken(token) // Save the token
             return token
         } else {
-            throw Exception("Failed to login user: ${response.status}")
+            throw Exception(response.bodyAsText())
         }
     }
 
     override suspend fun logoutUser() {
         AuthStorage.removeAuthToken()
+    }
+
+    override suspend fun getUserById(id: Int): User? {
+        val url = ServiceBuilder.url("api/users/$id")
+        val response: HttpResponse = client.get(url) {
+            contentType(ContentType.Application.Json)
+        }
+        return when (response.status) {
+            HttpStatusCode.OK -> {
+                val responseBody = response.bodyAsText()
+                Json.decodeFromString<User>(responseBody)
+            }
+            else -> null
+        }
+    }
+
+    private fun validateEmail(email: String) {
+        val emailRegex = "^[A-Za-z](.*)([@]{1})(.{1,})(\\.)(.{1,})"
+        if (!email.matches(emailRegex.toRegex())) {
+            throw RuntimeException("Invalid email format")
+        }
+    }
+
+    private fun validatePassword(password: String) {
+        if (password.length < 8 ||
+            !password.any { it.isUpperCase() } ||
+            !password.any { it.isLowerCase() } ||
+            !password.any { it.isDigit() } ||
+            !password.any { !it.isLetterOrDigit() }
+        ) {
+            throw RuntimeException("Password must be at least 8 characters long and include uppercase, lowercase, digit, and special character")
+        }
     }
 }