+++ PsycheAssistantAPI/src/main/java/com/example/PsycheAssistantAPI/Model/User.java
@@ -22,6 +22,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
+    @Column(nullable = false, unique = true)
     private String email;
     private String password;
     private int energyExpenditure = 0;