+++ PsycheAssistantAPI/src/main/java/com/example/PsycheAssistantAPI/Repository/GroupRepository.java
@@ -3,4 +3,6 @@ package com.example.PsycheAssistantAPI.Repository;
 import com.example.PsycheAssistantAPI.Model.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupRepository extends JpaRepository<Group, Integer> {}
+public interface GroupRepository extends JpaRepository<Group, Integer> {
+    Group findByCode(String code);
+}